cleanup integration tests with more descriptive names
