+++ oracle/tests/it/main.rs
@@ -2,6 +2,5 @@
 mod utils;
 
 mod init;
-mod dr_creation_tests;
-mod dr_resolution_tests;
+mod dr_basic_tests;
 mod dr_scenario_tests;
